+++ app/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [DemoApp][Library] Create component - Checkbox ([#47](https://github.com/Orange-OpenSource/ouds-flutter/issues/47))
 - [Library] `ControlItem` component (tokens library v0.11.0) ([#132](https://github.com/Orange-OpenSource/ouds-flutter/issues/132))
 - [DemoApp] Tokens: screen elevation ([#84](https://github.com/Orange-OpenSource/ouds-flutter/issues/84))
 - [Library] `Chekbox` component (tokens library v0.8.0) ([#120](https://github.com/Orange-OpenSource/ouds-flutter/issues/120))
@@ -30,7 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [DemoApp][Library] Update GitHub configuration files ([#106](https://github.com/Orange-OpenSource/ouds-flutter/issues/106))
-- [DemoApp] Create component - Button ([#45](https://github.com/Orange-OpenSource/ouds-flutter/issues/45))
+- [DemoApp][Library] Create component - Button ([#45](https://github.com/Orange-OpenSource/ouds-flutter/issues/45))
 - [DemoApp] About page correct spacing and add flutter label ([#80](https://github.com/Orange-OpenSource/ouds-flutter/issues/80))
 - [DemoApp] Update Flutter SDK to v3.6.0 ([#81](https://github.com/Orange-OpenSource/ouds-flutter/issues/81))
 - [Library] Fix workflow to use dart doc ([#75](https://github.com/Orange-OpenSource/ouds-flutter/issues/75))